Add files to allow testing without TripleO
Add a top-level playbook and a test-role/default/main.yml to allow running
without OoO.
The main.yml file is populated with the parameters added in from OoO
template which is WIP and unpublished.

The playbook can be run with::
  ansible-playbook -v default.inv main.yml

This will run with the default values used by OoO, and maines manual
testing/debugging easier for now.