[WIP][tripleo_collectd] Add role for tripleo
This roles adds the existing ansible roles that are embedded in
THT/collectd-container-ansible.yaml for deploy_steps_tasks,
fast_forward_upgrade_tasks and host_prep_tasks.

The tripleo_collectd role is untested, and requires additional steps and
logic.
At the moment, the vars are passed in from OoO, but need to be parsed to
replace some of the logic in THT.